django/requirements.txt
Django==5.1.1
django-cors-middleware==1.5.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-stubs==5.1.0
setuptools>=58.0.0
